Seperated Movies and TV_Show Models
