+++ app.py
@@ -4,7 +4,7 @@ from flask_cors import CORS
 import pymysql
 import os
 
-#Setup App
+# Setup App
 app = Flask(__name__)
 app.config.from_object(os.environ['APP_SETTINGS']) #Should change based on is in Development or Production
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
@@ -12,14 +12,15 @@ app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 # Enable CORS
 CORS(app)
 
-#Start Database
+# Start Database
 db = SQLAlchemy(app)
 
 # Enable Variable port for Heroku
 port = int(os.environ.get('PORT', 33507))
 
-#Import models
-from models import Actor, Movie, TV_Shows
+# Import models
+from models import Actor
+from media_models import Movie, TV_Shows
 
 # Force pymysql to be used as replacement for MySQLdb
 pymysql.install_as_MySQLdb()
